Face_information_model/requirements.txt
absl-py
appnope==0.1.0
astroid==2.4.2
astunparse==1.6.3
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
deepface==0.0.33

face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
gast==0.3.3
gdown==3.12.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio
h5py
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.6
jupyter-core==4.6.3


lazy-object-proxy
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1

oauthlib==3.1.0

opt-einsum==3.3.0
numpy
pandas
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit==3.0.5
protobuf
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.5.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML
pyzmq
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy
six==1.15.0
termcolor==1.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typed-ast
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0

